Add deserialization for query objects in QueryTests
This update introduces functionality to deserialize a serialized query object back into its original form. After serializing the `query` object to JSON, the code now uses `JsonSerializer.Deserialize` to convert the JSON string back into a `Query<Company, Company>` object. An assertion is also added to verify that the deserialized query is not null, ensuring the deserialization process is successful.